[NEW] Bookmark_database_manager class added [UPDATE] Load comif from bookmark manager dialog working
