+++ src/bookmark_manager_dialog.ui
@@ -9,8 +9,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>780</width>
-    <height>400</height>
+    <width>726</width>
+    <height>436</height>
    </rect>
   </property>
   <property name="sizePolicy">
@@ -82,12 +82,18 @@
       </widget>
      </item>
      <item>
-      <widget class="QWidget" name="verticalWidget" native="true">
+      <widget class="QFrame" name="verticalFrame">
        <property name="sizePolicy">
         <sizepolicy hsizetype="Minimum" vsizetype="Minimum">
          <horstretch>0</horstretch>
          <verstretch>0</verstretch>
         </sizepolicy>
+       </property>
+       <property name="frameShape">
+        <enum>QFrame::NoFrame</enum>
+       </property>
+       <property name="frameShadow">
+        <enum>QFrame::Raised</enum>
        </property>
        <layout class="QVBoxLayout" name="verticalLayout">
         <item>
@@ -96,7 +102,7 @@
            <enum>Qt::Vertical</enum>
           </property>
           <property name="sizeType">
-           <enum>QSizePolicy::Preferred</enum>
+           <enum>QSizePolicy::Expanding</enum>
           </property>
           <property name="sizeHint" stdset="0">
            <size>
@@ -112,7 +118,10 @@
            <string/>
           </property>
           <property name="pixmap">
-           <pixmap resource="../resource/main_window.qrc">:/icons/notCover.png</pixmap>
+           <pixmap resource="../resource/main_window.qrc">:/icons/pynocchio_icon_2.png</pixmap>
+          </property>
+          <property name="scaledContents">
+           <bool>true</bool>
           </property>
           <property name="alignment">
            <set>Qt::AlignCenter</set>
@@ -121,6 +130,9 @@
         </item>
         <item>
          <widget class="QLabel" name="page_preview_label">
+          <property name="enabled">
+           <bool>true</bool>
+          </property>
           <property name="font">
            <font>
             <italic>true</italic>
@@ -161,6 +173,17 @@
    <item>
     <layout class="QGridLayout" name="grid_layout">
      <item row="1" column="3">
+      <widget class="QPushButton" name="button_cancel">
+       <property name="text">
+        <string>Cancel</string>
+       </property>
+       <property name="icon">
+        <iconset resource="../resource/main_window.qrc">
+         <normaloff>:/icons/elementary3-icon-theme/actions/48/dialog-cancel.svg</normaloff>:/icons/elementary3-icon-theme/actions/48/dialog-cancel.svg</iconset>
+       </property>
+      </widget>
+     </item>
+     <item row="1" column="4">
       <widget class="QPushButton" name="button_load">
        <property name="text">
         <string>Load</string>
@@ -183,17 +206,6 @@
       </widget>
      </item>
      <item row="1" column="2">
-      <widget class="QPushButton" name="button_exit">
-       <property name="text">
-        <string>Close</string>
-       </property>
-       <property name="icon">
-        <iconset resource="../resource/main_window.qrc">
-         <normaloff>:/icons/elementary3-icon-theme/actions/48/dialog-cancel.svg</normaloff>:/icons/elementary3-icon-theme/actions/48/dialog-cancel.svg</iconset>
-       </property>
-      </widget>
-     </item>
-     <item row="1" column="0">
       <spacer name="horizontalSpacer">
        <property name="orientation">
         <enum>Qt::Horizontal</enum>
@@ -217,7 +229,7 @@
  </resources>
  <connections>
   <connection>
-   <sender>button_exit</sender>
+   <sender>button_cancel</sender>
    <signal>clicked()</signal>
    <receiver>Bookmark_Dialog</receiver>
    <slot>close()</slot>
